+++ console2/src/components/organisms/OrganizationOwnerChangeActivity/OrganizationOwnerChangeActivity.tsx
@@ -1,0 +1,82 @@
+/*-
+ * *****
+ * Concord
+ * -----
+ * Copyright (C) 2017 - 2018 Walmart Inc.
+ * -----
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =====
+ */
+
+import * as React from 'react';
+import { connect, Dispatch } from 'react-redux';
+
+import { ConcordId, ConcordKey, RequestError } from '../../../api/common';
+import { actions, State } from '../../../state/data/orgs';
+import { RequestErrorMessage } from '../../molecules';
+import EntityOwnerChangeForm from '../../molecules/EntityOwnerChangeForm';
+
+interface ExternalProps {
+    orgId: ConcordId;
+    orgName: ConcordKey;
+    owner?: string;
+}
+
+interface StateProps {
+    changing: boolean;
+    error: RequestError;
+}
+
+interface DispatchProps {
+    change: (orgId: ConcordId, orgName: ConcordKey, owner: string) => void;
+}
+
+type Props = ExternalProps & StateProps & DispatchProps;
+
+class OrganizationOwnerChangeActivity extends React.PureComponent<Props> {
+    constructor(props: Props) {
+        super(props);
+
+        this.state = { dirty: false, showConfirm: false, value: props.owner };
+    }
+
+    render() {
+        const { error, owner, changing, change, orgId, orgName } = this.props;
+
+        return (
+            <>
+                {error && <RequestErrorMessage error={error} />}
+                <EntityOwnerChangeForm
+                    originalOwner={owner || ''}
+                    confirmationHeader="Change organization owner?"
+                    confirmationContent="Are you sure you want to change the organization's owner?"
+                    onSubmit={(value) => change(orgId, orgName, value)}
+                    submitting={changing}
+                />
+            </>
+        );
+    }
+}
+
+const mapStateToProps = ({ orgs }: { orgs: State }): StateProps => ({
+    changing: orgs && orgs.changeOwner.running,
+    error: orgs && orgs.changeOwner.error
+});
+
+const mapDispatchToProps = (dispatch: Dispatch<{}>): DispatchProps => ({
+    change: (orgId, orgName, owner) => dispatch(actions.changeOwner(orgId, orgName, owner))
+});
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps
+)(OrganizationOwnerChangeActivity);